pyproject.toml
[project]
name = "opencode-webui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beam-client>=0.2.191",
]
